Move Future poll functions out of cxxbridge FFI module
This required somewhat of a large refactor. As part of this change, I made it a requirement that boilerplate Future types must return `Result<T, E>`, with infallible Futures now returning `Result<T, Infallible>`. This just makes our implementation a bit easier for now -- we could support non-`Result` return types at some point, but for now this made it easier to refactor.