Remove dependency on cxx::Exception
Using cxx::Exception was convenient when originally figuring out how to return exceptions from C++ to Rust outside of a cxx-rs bridge interface. However, it depends on private details of cxx-rs' implementation which are intended for cxx-rs' own generated code.

It'd be nice if cxx-rs were written in such a way that we could share an Error / Exception type with it. For now, though, I'm copying code in, as it's simple enough.